adding Models files, middleware, passport.js
hoping this works guys!
